Lazy solution to problem that doesn't exist locally
"baljan.RefillSeries.code_value: (fields.E004) 'choices' must be an
iterable (e.g., a list or tuple)." only appears when deploying to
Heroku. Even the django docs states that 'choices' can be a callable
function.

"choices can also be defined as a callable that expects no arguments and
returns any of the formats described above."