Update general code structure of repository (#57)
This squash commit contains a bunch of quality of life changes to the repository.

For example, I updated all dependencies, added a ruff config with a pre-commit hook to secure a unified code base, and renamed services etc to use the same name everywhere. I also moved away from a bunch of deprecated Heroku configurations.

* Format all files using new ruff config

* Fix is_ajax deprecation error

* Update Python version used by Docker

* Add pre-commit hook for ruff

* Add pre-commit hooks

* Update Celery implementation

* Remove unused files

* Update code_value choices

* Remove CELERY_IMPORTS field

* Update README